requirements.txt
google-adk>=1.19.0
google-cloud-aiplatform[adk,agent-engines]>=1.100.0
google-cloud-storage>=3.0.0
pydantic>=2.10.0
python-dotenv>=1.1.1
absl-py>=2.2.1
pytest>=8.3.5
pytest-asyncio>=0.26.0

